Implement DI scopes
... and reuse them for Sisu
