+++ maven-core/src/test/java/org/apache/maven/execution/scope/internal/MojoExecutionScopeTest.java
@@ -39,15 +39,15 @@ class MojoExecutionScopeTest {
 
         Object o1 = new Object();
         scope.seed(Object.class, o1);
-        assertSame(o1, scope.scope(Key.get(Object.class), null).get());
+        assertSame(o1, scope.scope(Key.get(Object.class), () -> null).get());
 
         scope.enter();
         Object o2 = new Object();
         scope.seed(Object.class, o2);
-        assertSame(o2, scope.scope(Key.get(Object.class), null).get());
+        assertSame(o2, scope.scope(Key.get(Object.class), () -> null).get());
 
         scope.exit();
-        assertSame(o1, scope.scope(Key.get(Object.class), null).get());
+        assertSame(o1, scope.scope(Key.get(Object.class), () -> null).get());
 
         scope.exit();
